refactor(core): split session-manager.ts into focused modules
Extract OpenCode-specific helpers, PR claim, and session restore/remap
into dedicated files sharing a SessionManagerContext dependency bundle.
Public API of createSessionManager() is unchanged.

Closes #1415
